plugin/client/requirements.txt
pydirectinput
python-socketio[asyncio_client]==4.6.1
pywin32
pillow
rich
